+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 # Change Log
+## [v0.5.0] - 2022-02-08
+Relatively signficant update: Layering system updated to draw all elements in one group and use a global z-index, instead of many separate local z-indexes that elements will switch to.
+### Added
+* Infographic elements are now all drawn in main group, instead of being grouped into subgroups
+### Changed
+* PositionCommand updated to not use LayerCommand
+### Removed
+* _SwitchContainerOnDrag in AInfographic
 ## [v0.4.1] - 2022-02-07
 ### Changed
 * ClearSelection is now called before text elements are inserted instead of afterwards